API: Save codes internally and organize by length
The catalog of best codes is now save in a module internal to jeweler,
so that the work for the codes can be done for the end user and
distributed. Also, codes are now organized by length so that there are
many small files instead of one large file. This is better because we
may want to generate codes of different weights for various noise types.